Move Tune annotations to models module
