+++ folk_rnn_site/archiver/views.py
@@ -6,8 +6,7 @@ from django.utils.timezone import now
 from django.core.exceptions import ValidationError
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.contrib.postgres.search import SearchQuery, SearchRank, SearchVector
-from django.db.models import F, Q, Count
-from django.db.models.query import QuerySet
+from django.db.models import Q
 from tempfile import TemporaryFile
 from itertools import chain
 from datetime import timedelta
@@ -16,7 +15,7 @@ from random import choice, choices
 from folk_rnn_site.models import ABCModel
 from archiver import TUNE_SEARCH_EXAMPLES, MAX_RECENT_ITEMS, TUNE_PREVIEWS_PER_PAGE
 from archiver import weightedSelectionWithoutReplacement
-from archiver.models import User, Tune, TuneAttribution, Setting, Comment, Recording, Event, TunebookEntry, TuneRecording
+from archiver.models import User, Tune, annotate_counts, TuneAttribution, Setting, Comment, Recording, Event, TunebookEntry, TuneRecording
 from archiver.forms import (
                             SettingForm, 
                             CommentForm, 
@@ -29,30 +28,6 @@ from archiver.forms import (
                             SearchForm,
                             )
 from archiver.dataset import dataset_as_csv
-
-def queryset_annotate_counts(self):
-    return self.annotate(
-        Count('setting', distinct=True), 
-        Count('comment', distinct=True), 
-        recording__count=Count('tunerecording', distinct=True), 
-        event__count=Count('tuneevent', distinct=True),
-        tunebook__count=Count('tunebookentry', distinct=True),
-    )
-setattr(QuerySet, 'annotate_counts', queryset_annotate_counts)
-
-def annotate_counts(tunes):
-    for result in tunes:
-        result.setting__count = result.setting_set.count()
-        result.comment__count = result.comment_set.count() 
-        result.recording__count = result.tunerecording_set.count() 
-        result.event__count = result.tuneevent_set.count()
-        result.tunebook__count = result.tunebookentry_set.count()
-
-def queryset_annotate_saliency(self):
-    return self.annotate(
-        saliency = F('tunebook__count')*3 + F('setting__count')*3 + F('recording__count')*2 + F('event__count')*2 + F('comment__count')
-    )
-setattr(QuerySet, 'annotate_saliency', queryset_annotate_saliency)
 
 def activity(filter_dict={}):
     qs_tune = Tune.objects.filter(**filter_dict).order_by('-id')[:MAX_RECENT_ITEMS]
